+++ services/web/server/tests/unit/isolated/test_garbage_collector_core.py
@@ -10,13 +10,15 @@ from aiohttp import ClientSession, web
 from aioresponses import aioresponses as AioResponsesMock
 from faker import Faker
 from pytest_mock import MockerFixture
-from simcore_service_webserver.garbage_collector._core import (
-    _remove_orphaned_services,
+from simcore_service_webserver.garbage_collector._core_orphans import (
     _remove_single_service_if_orphan,
+    remove_orphaned_services,
 )
 from yarl import URL
 
-MODULE_GC_CORE: Final[str] = "simcore_service_webserver.garbage_collector._core"
+MODULE_GC_CORE_ORPHANS: Final[
+    str
+] = "simcore_service_webserver.garbage_collector._core_orphans"
 
 
 @pytest.fixture
@@ -37,7 +39,7 @@ def mock_get_workbench_node_ids_from_project_uuid(
     mocker: MockerFixture, faker: Faker
 ) -> None:
     mocker.patch(
-        f"{MODULE_GC_CORE}.get_workbench_node_ids_from_project_uuid",
+        f"{MODULE_GC_CORE_ORPHANS}.get_workbench_node_ids_from_project_uuid",
         return_value={faker.uuid4(), faker.uuid4(), faker.uuid4()},
     )
 
@@ -45,18 +47,18 @@ def mock_get_workbench_node_ids_from_project_uuid(
 @pytest.fixture
 def mock_list_dynamic_services(mocker: MockerFixture):
     mocker.patch(
-        f"{MODULE_GC_CORE}.api.list_dynamic_services",
+        f"{MODULE_GC_CORE_ORPHANS}.api.list_dynamic_services",
         autospec=True,
     )
 
 
-async def test_regression_remove_orphaned_services_node_ids_unhashable_type_set(
+async def test_regressionremove_orphaned_services_node_ids_unhashable_type_set(
     mock_get_workbench_node_ids_from_project_uuid: None,
     mock_list_dynamic_services: None,
     mock_registry: AsyncMock,
     mock_app: AsyncMock,
 ):
-    await _remove_orphaned_services(mock_registry, mock_app)
+    await remove_orphaned_services(mock_registry, mock_app)
 
 
 async def test_regression_project_id_recovered_from_the_wrong_data_structure(
@@ -65,16 +67,16 @@ async def test_regression_project_id_recovered_from_the_wrong_data_structure(
     # tests that KeyError is not raised
 
     mocker.patch(
-        f"{MODULE_GC_CORE}.is_node_id_present_in_any_project_workbench",
+        f"{MODULE_GC_CORE_ORPHANS}.is_node_id_present_in_any_project_workbench",
         autospec=True,
         return_value=True,
     )
     mocker.patch(
-        f"{MODULE_GC_CORE}.ProjectDBAPI.get_from_app_context",
+        f"{MODULE_GC_CORE_ORPHANS}.ProjectDBAPI.get_from_app_context",
         autospec=True,
         return_value=AsyncMock(),
     )
-    mocker.patch(f"{MODULE_GC_CORE}.api.stop_dynamic_service", autospec=True)
+    mocker.patch(f"{MODULE_GC_CORE_ORPHANS}.api.stop_dynamic_service", autospec=True)
 
     await _remove_single_service_if_orphan(
         app=AsyncMock(),
@@ -94,12 +96,12 @@ async def test_remove_single_service_if_orphan_service_is_waiting_manual_interve
     aioresponses_mocker: AioResponsesMock,
 ):
     mocker.patch(
-        f"{MODULE_GC_CORE}.is_node_id_present_in_any_project_workbench",
+        f"{MODULE_GC_CORE_ORPHANS}.is_node_id_present_in_any_project_workbench",
         autospec=True,
         return_value=True,
     )
     mocker.patch(
-        f"{MODULE_GC_CORE}.ProjectDBAPI.get_from_app_context",
+        f"{MODULE_GC_CORE_ORPHANS}.ProjectDBAPI.get_from_app_context",
         autospec=True,
         return_value=AsyncMock(),
     )
